process_task_api/requirements.txt
flask
requests
pandas
numpy
pillow
google-cloud-storage
redis 
google-cloud-pubsub